3.2

+++ script.sh
@@ -1,25 +1,55 @@
 #!/bin/sh
 
-GITHUB_REPOSITORY=https://github.com/jylhakos/phonebook.git
+GITHUB_USER="jylhakos"
 
-DOCKER_REPOSITORY=phonebook
+GITHUB_TOKEN=""
 
-DIRECTORY=tmp
+GITHUB_REPOSITORY="github.com/jylhakos/phonebook.git"
 
-IMAGE=phonebook
+REPOSITORY_DIRECTORY="phonebook"
 
-rm -rf $DIRECTORY
+GITHUB_REPOSITORY_URL="https://"$GITHUB_REPOSITORY
+
+DOCKER_USER="jylhakos"
+
+DOCKER_TOKEN=""
+
+DOCKER_IMAGE="phonebook"
+
+DIRECTORY="exercise"
+
+HAS_DOCKER_IMAGE=$( docker images -q $DOCKER_IMAGE )
+
+echo "HAS_DOCKER_IMAGE" $HAS_DOCKER_IMAGE
+
+if [[ -n "$HAS_DOCKER_IMAGE" ]] 
+then
+	echo "remove $DOCKER_IMAGE"
+
+	docker rmi -f DOCKER_IMAGE
+fi
+
+if [[ -n "$DIRECTORY" ]]
+then
+	echo "remove $DIRECTORY"
+
+	rm -rf $DIRECTORY
+fi
 
 mkdir $DIRECTORY
 
-cd $DIRECTORY
+cd $DIRECTORY"/"$REPOSITORY_DIRECTORY
 
-git clone $GITHUB_REPOSITORY
+pwd
+
+git clone $GITHUB_REPOSITORY_URL
 
 npm install
 
-docker login
+ls -l
 
-docker build -t $IMAGE .
+docker build -t $DOCKER_USER"/"$DOCKER_IMAGE .
 
-docker push jylhakos/$DOCKER_REPOSITORY
+docker login --username DOCKER_USER --password DOCKER_TOKEN
+
+docker push $DOCKER_USER"/"$DOCKER_IMAGE
